Kolejne zmiany w pobieraniu danych
