+++ src/db/Driver.java
@@ -1,5 +1,13 @@
 package db;
 
+import javafx.beans.property.SimpleStringProperty;
+import javafx.beans.value.ObservableValue;
+import javafx.collections.FXCollections;
+import javafx.collections.ObservableList;
+import javafx.scene.control.TableColumn;
+import javafx.scene.control.TableView;
+import javafx.util.Callback;
+
 import java.sql.*;
 
 public class Driver {
@@ -77,4 +85,41 @@ public class Driver {
             }
         }
     }
+
+    public static TableView getResult(String query) {
+        TableView tableview = new TableView();
+        Statement statement;
+        ObservableList<ObservableList> data = FXCollections.observableArrayList();
+        try{
+            statement = Driver.getConnection().createStatement();
+            ResultSet rs = statement.executeQuery(query);
+
+            for(int i = 0; i < rs.getMetaData().getColumnCount(); ++i){
+                final int j = i;
+                TableColumn col = new TableColumn(rs.getMetaData().getColumnName(i+1));
+                col.setCellValueFactory(new Callback<TableColumn.CellDataFeatures<ObservableList,String>, ObservableValue<String>>(){
+                    public ObservableValue<String> call(TableColumn.CellDataFeatures<ObservableList, String> param) {
+                        return new SimpleStringProperty(param.getValue().get(j).toString());
+                    }
+                });
+
+                tableview.getColumns().addAll(col);
+            }
+
+            while(rs.next()){
+                ObservableList<String> row = FXCollections.observableArrayList();
+                for(int i = 1 ; i <= rs.getMetaData().getColumnCount(); ++i){
+                    row.add(rs.getString(i));
+                }
+                data.add(row);
+            }
+            tableview.setItems(data);
+        } catch(Exception e){
+            e.printStackTrace();
+            System.out.println("Error on Building Data");
+        }
+
+        return tableview;
+    }
+
 }
